+++ src/main/kotlin/com/dobby/backend/DobbyBackendApplication.kt
@@ -1,9 +1,14 @@
 package com.dobby.backend
 
 import com.dobby.backend.application.usecase.UseCase
+import com.dobby.backend.infrastructure.config.properties.GoogleAuthProperties
+import com.dobby.backend.infrastructure.config.properties.NaverAuthProperties
+import com.dobby.backend.infrastructure.config.properties.S3Properties
+import com.dobby.backend.infrastructure.config.properties.TokenProperties
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan
+import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.cloud.openfeign.EnableFeignClients
 import org.springframework.context.annotation.ComponentScan
 import org.springframework.context.annotation.FilterType
@@ -17,6 +22,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing
 )
 @SpringBootApplication
 @ConfigurationPropertiesScan
+@EnableConfigurationProperties(TokenProperties::class, S3Properties::class, GoogleAuthProperties::class, NaverAuthProperties::class)
 @EnableFeignClients
 @EnableJpaAuditing
 class DobbyBackendApplication
