+++ src/main/kotlin/com/dobby/backend/infrastructure/ContentLengthRequestInterceptor.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.config
+package com.dobby.backend.infrastructure
 
 import feign.RequestInterceptor
 import feign.RequestTemplate
@@ -11,7 +11,7 @@ class ContentLengthRequestInterceptor : RequestInterceptor {
         val length = if (requestTemplate.body() != null) requestTemplate.body().size else 0
         if (length == 0 && requestTemplate.method() == "POST") {
             requestTemplate.body("gradmeet")
-            requestTemplate.header(CONTENT_LENGTH_HEADER, "hello".toByteArray().size.toString())
+            requestTemplate.header(CONTENT_LENGTH_HEADER, "gradmeet".toByteArray().size.toString())
         }
     }
